UT-3577: move run_siegfried to utilities and write some more pydoc
